fix: use `Plugin` from the commitlint types
This makes it easier to maintain and is a lot nicer as it's using the type the plugin expects rather than a redefinition of it.

As the rules are specified separately from where `Plugin` is used, the rules use `Plugin['rules']` as their type. The `as Rule<unknown>` part of the rule definitions is required as that is the type commitlint is expecting. It'd be better if they just require `Rule` as then this cast isn't necessary. I'm making a PR for that in the commitlint repo (https://github.com/conventional-changelog/commitlint).